training-job/requirements.txt
numpy
pandas
sklearn
pytorch_lightning
transformers
torchdata
mosaicml
